--- FASTAPI/app/schemas/CategorySchema.py
@@ -1,19 +1,0 @@
-from pydantic import BaseModel
-from typing import Optional
-
-class CategoryBase(BaseModel):
-    name: str
-    description: Optional[str] = None
-    parent_category_id: Optional[int] = None
-
-class CategoryCreate(CategoryBase):
-    pass
-
-class CategoryUpdate(CategoryBase):
-    pass
-
-class CategoryResponse(CategoryBase):
-    category_id: int
-
-    class Config:
-        from_attributes = True
