--- FASTAPI/app/schemas/UserSchema.py
@@ -1,25 +1,0 @@
-from pydantic import BaseModel, EmailStr
-from typing import Optional
-from datetime import datetime
-
-class UserBase(BaseModel):
-    username: str
-    password_hash: str
-    full_name: str
-    email: Optional[EmailStr] = None
-    role_id: int
-    is_active: Optional[bool] = True
-    created_at: Optional[datetime] = None
-    updated_at: Optional[datetime] = None
-
-class UserCreate(UserBase):
-    pass
-
-class UserUpdate(UserBase):
-    pass
-
-class UserResponse(UserBase):
-    user_id: int
-
-    class Config:
-        from_attributes = True
